+++ x/supply/client/cli/query.go
@@ -17,7 +17,7 @@ import (
 func GetQueryCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:                        types.ModuleName,
-		Short:                      "Querying commands for the supply module",
+		Short:                      "Querying commands for the supply module [Deprecated: do not use]",
 		DisableFlagParsing:         true,
 		SuggestionsMinimumDistance: 2,
 		RunE:                       client.ValidateCmd,
@@ -35,9 +35,9 @@ func GetQueryCmd() *cobra.Command {
 func GetCmdQueryTotalSupply() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "total",
-		Short: "Query the total supply of coins of the chain",
+		Short: "Query the total supply of coins of the chain [Deprecated: do not use]",
 		Long: strings.TrimSpace(
-			fmt.Sprintf(`Query total supply of coins that are held by accounts in the chain.
+			fmt.Sprintf(`Query total supply of coins that are held by accounts in the chain. [Deprecated: do not use]
 Example:
   $ %s query %s total
 `,
@@ -50,6 +50,7 @@ Example:
 				return err
 			}
 
+			// nolint: staticcheck
 			queryClient := types.NewQueryClient(clientCtx)
 			res, err := queryClient.TotalSupply(cmd.Context(), types.NewSupplyRequest())
 			if err != nil {
@@ -69,9 +70,9 @@ Example:
 func GetCmdQueryLiquidSupply() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "liquid",
-		Short: "Query the liquid supply of coins of the chain",
+		Short: "Query the liquid supply of coins of the chain [Deprecated: do not use]",
 		Long: strings.TrimSpace(
-			fmt.Sprintf(`Query liquid supply of coins that are held by accounts in the chain.
+			fmt.Sprintf(`Query liquid supply of coins that are held by accounts in the chain. [Deprecated: do not use]
 Example:
   $ %s query %s liquid
 `,
@@ -84,6 +85,7 @@ Example:
 				return err
 			}
 
+			// nolint: staticcheck
 			queryClient := types.NewQueryClient(clientCtx)
 			res, err := queryClient.LiquidSupply(cmd.Context(), types.NewSupplyRequest())
 			if err != nil {